requirements.txt
Flask==2.0.1
pymongo==3.12.0
flask-pymongo==2.3.0
werkzeug==2.0.3
pymongo[srv]
Flask-JWT-Extended
python-dotenv
waitress
